Merge remote-tracking branch 'origin/master'
Add project.html page: The page serves to display offers by the user.
After adding the page, with the display of all offers, the user can choose a specific offer, then by clicking on it, the detailed content of the ad will be displayed with more information of that ad.

Also you need to add  in view.py file, function def OfferView() and def ShowOffer(), which includes model from the base on the page.

For ShowOffer function, we need to set context to get are Offer ID number, to show detail information of that specific post.

You can also add some API context to your dashboard.html page:
In my example I add API for weather (for city of Belgrade)
For this example I am using JSON library.